Fixed bugs for 0^0 any location of the string and first term powers
Fix 1: 0^0 and 0/0 would be skipped over if the expression was something like 6*0+0^0+6*0 or 6*0+0/0+6*0, this has now been resolved to return the appropriate error using conversion in different areas. 
Fix 2: The first term of every string when using powers and a negative sign is present would perform the calculation on the whole term e.g. currently the calculator would perform -2^2 as (-2)^2, this is incorrect until parenthesis are enabled. Now -2^2 would be resolved as -(2)^2 provided its the first term of the string.